WIP Interface refactor
