+++ apps/ui/src/lib/components/ContextMenu.svelte
@@ -1,3 +1,4 @@
-<nav class="header px-0 mb-5" style="border-bottom: 2px solid #666;">
+<nav class="header justify-between px-0 mb-5" style="border-bottom: 2px solid #666;">
   <slot />
+  <slot name="actions"></slot>
 </nav>
