+++ apps/ui/src/lib/components/MainNavigationNew.svelte
@@ -3,9 +3,18 @@
   import { appSession } from '$lib/store';
   import Tooltip from '$lib/components/Tooltip.svelte';
   import UpdateAvailable from '$lib/components/UpdateAvailable.svelte';
+	import DashboardIcon from './svg/menu/DashboardIcon.svelte';
+	import ServersIcon from './svg/menu/ServersIcon.svelte';
+	import DatabasesIcon from './svg/menu/DatabasesIcon.svelte';
+	import IamIcon from './svg/menu/IamIcon.svelte';
+	import SettingsIcon from './svg/menu/SettingsIcon.svelte';
+	import LogoutIcon from './svg/menu/LogoutIcon.svelte';
+	import SourcesIcon from './svg/menu/SourcesIcon.svelte';
 </script>
+
 {#if $appSession.userId}
   <nav class="hidden lg:flex flex-row " style="background: rgb(32 32 32 / var(--tw-bg-opacity)); border-bottom: thin solid #8884">
+    <!-- Logo vs Whitelabel -->
     {#if !$appSession.whiteLabeled}
       <div class="m-2 h-10 w-10">
         <img src="/favicon.png" alt="coolLabs logo" />
@@ -15,6 +24,8 @@
         <img src={$appSession.whiteLabeledDetails.icon} alt="White labeled logo" />
       </div>
     {/if}
+
+    <!-- Left menu -->
     <div class="flex flex-row" class:mt-2={$appSession.whiteLabeled}>
       <a
         id="dashboard"
@@ -25,22 +36,18 @@
         class:bg-coolgray-500={$page.url.pathname === '/'}
         class:bg-coolgray-200={!($page.url.pathname === '/')}
       >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          class="m-1 h-9 w-9"
-          viewBox="0 0 24 24"
-          stroke-width="1.5"
-          stroke="currentColor"
-          fill="none"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-          <path
-            d="M19 8.71l-5.333 -4.148a2.666 2.666 0 0 0 -3.274 0l-5.334 4.148a2.665 2.665 0 0 0 -1.029 2.105v7.2a2 2 0 0 0 2 2h12a2 2 0 0 0 2 -2v-7.2c0 -.823 -.38 -1.6 -1.03 -2.105"
-          />
-          <path d="M16 15c-2.21 1.333 -5.792 1.333 -8 0" />
-        </svg>
+       <DashboardIcon/>
+      </a>
+      <a
+        id="sources"
+        sveltekit:prefetch
+        href="/sources"
+        class="icons hover:text-sky-500"
+        class:text-sky-500={$page.url.pathname === '/sources'}
+        class:bg-coolgray-500={$page.url.pathname === '/sources'}
+        class:bg-coolgray-200={!($page.url.pathname === '/sources')}
+      >
+        <SourcesIcon/>
       </a>
       {#if $appSession.teamId === '0'}
         <a
@@ -52,31 +59,32 @@
           class:bg-coolgray-500={$page.url.pathname === '/servers'}
           class:bg-coolgray-200={!($page.url.pathname === '/servers')}
         >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            class="m-1 h-9 w-9"
-            viewBox="0 0 24 24"
-            stroke-width="1.5"
-            stroke="currentColor"
-            fill="none"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-          >
-            <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-            <rect x="3" y="4" width="18" height="8" rx="3" />
-            <rect x="3" y="12" width="18" height="8" rx="3" />
-            <line x1="7" y1="8" x2="7" y2="8.01" />
-            <line x1="7" y1="16" x2="7" y2="16.01" />
-          </svg>
+          <ServersIcon/>
         </a>
+        <a
+        id="databases"
+        sveltekit:prefetch
+        href="/databases"
+        class="icons hover:text-sky-500"
+        class:text-sky-500={$page.url.pathname === '/databases'}
+        class:bg-coolgray-500={$page.url.pathname === '/databases'}
+        class:bg-coolgray-200={!($page.url.pathname === '/databases')}
+      >
+        <DatabasesIcon/>
+      </a>
       {/if}
     </div>
     <Tooltip triggeredBy="#dashboard" placement="right">Dashboard</Tooltip>
     <Tooltip triggeredBy="#servers" placement="right">Servers</Tooltip>
+    <Tooltip triggeredBy="#databases" placement="right">Databases</Tooltip>
+    <Tooltip triggeredBy="#sources" placement="right">Sources</Tooltip>
+    
     <div class="flex-1" />
     <div class="lg:block hidden">
       <UpdateAvailable />
     </div>
+
+    <!-- Right menu -->
     <div class="flex flex-row">
       <a
         id="iam"
@@ -90,22 +98,8 @@
           <span class="indicator-item rounded-full badge badge-primary mr-2"
             >{pendingInvitations.length}</span
           >
-        {/if}<svg
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 24 24"
-          class="m-1 h-9 w-9"
-          stroke-width="1.5"
-          stroke="currentColor"
-          fill="none"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-          <circle cx="9" cy="7" r="4" />
-          <path d="M3 21v-2a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v2" />
-          <path d="M16 3.13a4 4 0 0 1 0 7.75" />
-          <path d="M21 21v-2a4 4 0 0 0 -3 -3.85" />
-        </svg>
+        {/if}
+        <IamIcon/>
       </a>
       <a
         id="settings"
@@ -115,22 +109,7 @@
         class:text-settings={$page.url.pathname.startsWith('/settings')}
         class:bg-coolgray-500={$page.url.pathname.startsWith('/settings')}
       >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 24 24"
-          class="m-1 h-9 w-9"
-          stroke-width="1.5"
-          stroke="currentColor"
-          fill="none"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-          <path
-            d="M10.325 4.317c.426 -1.756 2.924 -1.756 3.35 0a1.724 1.724 0 0 0 2.573 1.066c1.543 -.94 3.31 .826 2.37 2.37a1.724 1.724 0 0 0 1.065 2.572c1.756 .426 1.756 2.924 0 3.35a1.724 1.724 0 0 0 -1.066 2.573c.94 1.543 -.826 3.31 -2.37 2.37a1.724 1.724 0 0 0 -2.572 1.065c-.426 1.756 -2.924 1.756 -3.35 0a1.724 1.724 0 0 0 -2.573 -1.066c-1.543 .94 -3.31 -.826 -2.37 -2.37a1.724 1.724 0 0 0 -1.065 -2.572c-1.756 -.426 -1.756 -2.924 0 -3.35a1.724 1.724 0 0 0 1.066 -2.573c-.94 -1.543 .826 -3.31 2.37 -2.37c1 .608 2.296 .07 2.572 -1.065z"
-          />
-          <circle cx="12" cy="12" r="3" />
-        </svg>
+        <SettingsIcon/>
       </a>
 
       <!-- svelte-ignore a11y-click-events-have-key-events -->
@@ -139,22 +118,7 @@
         class="icons bg-coolgray-200 hover:text-error cursor-pointer"
         on:click={logout}
       >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          class="m-1 h-9 w-9"
-          viewBox="0 0 24 24"
-          stroke-width="1.5"
-          stroke="currentColor"
-          fill="none"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <path stroke="none" d="M0 0h24v24H0z" fill="none" />
-          <path
-            d="M14 8v-2a2 2 0 0 0 -2 -2h-7a2 2 0 0 0 -2 2v12a2 2 0 0 0 2 2h7a2 2 0 0 0 2 -2v-2"
-          />
-          <path d="M7 12h14l-3 -3m0 6l3 -3" />
-        </svg>
+        <LogoutIcon/>
       </div>
       <div
         class="m-1 mr-3 self-center font-bold text-stone-400 hover:bg-coolgray-200 hover:text-white"
