+++ apps/ui/src/lib/components/RightSidebar.svelte
@@ -1,0 +1,10 @@
+<div class="flex flex-row">
+  <div class="basis-2/3">
+    <slot name="content"/>
+  </div>
+  <div class="basis-1/3">
+    <div class="ml-4">
+      <slot name="sidebar"/>
+    </div>
+  </div>
+</div>
