+++ apps/ui/src/lib/components/svg/menu/ServersIcon.svelte
@@ -1,0 +1,16 @@
+<svg
+  xmlns="http://www.w3.org/2000/svg"
+  class="m-1 h-9 w-9"
+  viewBox="0 0 24 24"
+  stroke-width="1.5"
+  stroke="currentColor"
+  fill="none"
+  stroke-linecap="round"
+  stroke-linejoin="round"
+>
+  <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+  <rect x="3" y="4" width="18" height="8" rx="3" />
+  <rect x="3" y="12" width="18" height="8" rx="3" />
+  <line x1="7" y1="8" x2="7" y2="8.01" />
+  <line x1="7" y1="16" x2="7" y2="16.01" />
+</svg>
