+++ apps/ui/src/lib/screens/AppsNothingFound.svelte
@@ -1,0 +1,8 @@
+<script>
+  export let searched
+</script>
+<div class="flex flex-col items-center justify-center h-full pt-20">
+  <h1 class="text-2xl font-bold pb-4">
+    Nothing found with <span class="text-error font-bold">{searched}</span>.
+  </h1>
+</div>
