+++ apps/ui/src/routes/_NewResource.svelte
@@ -32,7 +32,7 @@
 					stroke-width="2"
 					d="M12 6v6m0 0v6m0-6h6m-6 0H6"
 				/></svg
-			> Create New Resource
+			> Create Resource
 		</label>
 	</slot>
