+++ apps/ui/src/routes/iam/teams.svelte
@@ -25,6 +25,7 @@
 	import { appSession } from '$lib/store';
 	import { errorNotification } from '$lib/common';
 	import { goto } from '$app/navigation';
+	import ContextMenu from '$lib/components/ContextMenu.svelte';
 
 	async function switchTeam(selectedTeamId: any) {
 		try {
@@ -50,14 +51,13 @@
 	}
 </script>
 
-<div class="w-full">
-	<div class="mx-auto w-full">
-		<div class="flex flex-row border-b border-coolgray-500 mb-6 space-x-2 items-center  pb-3">
-			<div class="title font-bold">Teams</div>
-			<button on:click={newTeam} class="btn btn-sm btn-primary"> Add New Team </button>
-		</div>
+<ContextMenu>
+	<div class="subtitle">Teams</div>
+	<div slot="actions">
+		<button on:click={newTeam} class="btn btn-sm btn-primary"> Add New Team </button>
 	</div>
-</div>
+</ContextMenu>
+
 <div class="grid grid-col gap-4 auto-cols-max grid-cols-1 md:grid-cols-2 lg:grid-cols-2 px-6">
 	{#each ownTeams as team}
 		<a href="/iam/teams/{team.id}" class="p-2 no-underline">
