+++ apps/ui/src/routes/sources/index.svelte
@@ -1,4 +1,164 @@
-<script lang="ts">
-	import { goto } from '$app/navigation';
-	goto('/');
+<script context="module" lang="ts">
+	import type { Load } from '@sveltejs/kit';
+	import {loadResources} from '$lib/resources';
+	export const load: Load = loadResources;
 </script>
+
+<script lang="ts">
+	export let gitSources:any;
+	import ContextMenu from "$lib/components/ContextMenu.svelte";
+	import GithubIcon from "$lib/components/svg/sources/GithubIcon.svelte";
+	import GitlabIcon from "$lib/components/svg/sources/GitlabIcon.svelte";
+</script>
+
+
+<ContextMenu>
+	<div class="title">Git Sources</div>
+</ContextMenu>
+
+{#if (gitSources.length > 0 && gitSources.length > 0) || gitSources.length > 0}
+		<div class="flex items-center mt-10">
+			<h1 class="title lg:text-3xl">Git Sources</h1>
+		</div>
+	{/if}
+	{#if gitSources.length > 0 && gitSources.length > 0}
+		<div class="divider" />
+		<div
+			class="grid grid-col gap-8 auto-cols-max grid-cols-1 md:grid-cols-2 lg:md:grid-cols-3 xl:grid-cols-4 p-4"
+		>
+			{#if gitSources.length > 0}
+				{#each gitSources as source}
+					<a class="no-underline mb-5" href={`/sources/${source.id}`}>
+						<div class="w-full rounded p-5 bg-coolgray-200 hover:bg-sources indicator duration-150">
+							<div class="w-full flex flex-row">
+								<div class="absolute top-0 left-0 -m-5 flex">
+									{#if source?.type === 'gitlab'}
+										<GitlabIcon />
+									{:else if source?.type === 'github'}
+										<GithubIcon/>
+									{/if}
+
+									{#if source.isSystemWide}
+										<svg
+											xmlns="http://www.w3.org/2000/svg"
+											class="h-10 w-10"
+											viewBox="0 0 24 24"
+											stroke-width="1.5"
+											stroke="currentColor"
+											fill="none"
+											stroke-linecap="round"
+											stroke-linejoin="round"
+										>
+											<path stroke="none" d="M0 0h24v24H0z" fill="none" />
+											<circle cx="12" cy="12" r="9" />
+											<line x1="3.6" y1="9" x2="20.4" y2="9" />
+											<line x1="3.6" y1="15" x2="20.4" y2="15" />
+											<path d="M11.5 3a17 17 0 0 0 0 18" />
+											<path d="M12.5 3a17 17 0 0 1 0 18" />
+										</svg>
+									{/if}
+								</div>
+								<div class="w-full flex flex-col">
+									<div class="h-10">
+										<h1 class="font-bold text-base truncate">{source.name}</h1>
+										{#if source.teams.length > 0 && source.teams[0]?.name}
+											<div class="truncate text-xs">{source.teams[0]?.name}</div>
+										{/if}
+									</div>
+
+									<div class="flex justify-end items-end space-x-2 h-10" />
+								</div>
+							</div>
+						</div>
+					</a>
+				{/each}
+			{:else}
+				<h1 class="">Nothing here.</h1>
+			{/if}
+		</div>
+	{/if}
+	{#if gitSources.length > 0}
+		{#if gitSources.length > 0}
+			<div class="divider w-32 mx-auto" />
+		{/if}
+	{/if}
+	{#if gitSources.length > 0}
+		<div
+			class="grid grid-col gap-8 auto-cols-max grid-cols-1 md:grid-cols-2 lg:md:grid-cols-3 xl:grid-cols-4 p-4"
+		>
+			{#each gitSources as source}
+				<a class="no-underline mb-5" href={`/sources/${source.id}`}>
+					<div class="w-full rounded p-5 bg-coolgray-200 hover:bg-sources indicator duration-150">
+						<div class="w-full flex flex-row">
+							<div class="absolute top-0 left-0 -m-5 flex">
+								{#if source?.type === 'gitlab'}
+									<svg viewBox="0 0 128 128" class="h-10 w-10">
+										<path
+											fill="#FC6D26"
+											d="M126.615 72.31l-7.034-21.647L105.64 7.76c-.716-2.206-3.84-2.206-4.556 0l-13.94 42.903H40.856L26.916 7.76c-.717-2.206-3.84-2.206-4.557 0L8.42 50.664 1.385 72.31a4.792 4.792 0 001.74 5.358L64 121.894l60.874-44.227a4.793 4.793 0 001.74-5.357"
+										/><path fill="#E24329" d="M64 121.894l23.144-71.23H40.856L64 121.893z" /><path
+											fill="#FC6D26"
+											d="M64 121.894l-23.144-71.23H8.42L64 121.893z"
+										/><path
+											fill="#FCA326"
+											d="M8.42 50.663L1.384 72.31a4.79 4.79 0 001.74 5.357L64 121.894 8.42 50.664z"
+										/><path
+											fill="#E24329"
+											d="M8.42 50.663h32.436L26.916 7.76c-.717-2.206-3.84-2.206-4.557 0L8.42 50.664z"
+										/><path fill="#FC6D26" d="M64 121.894l23.144-71.23h32.437L64 121.893z" /><path
+											fill="#FCA326"
+											d="M119.58 50.663l7.035 21.647a4.79 4.79 0 01-1.74 5.357L64 121.894l55.58-71.23z"
+										/><path
+											fill="#E24329"
+											d="M119.58 50.663H87.145l13.94-42.902c.717-2.206 3.84-2.206 4.557 0l13.94 42.903z"
+										/>
+									</svg>
+								{:else if source?.type === 'github'}
+									<svg viewBox="0 0 128 128" class="h-10 w-10">
+										<g fill="#ffffff"
+											><path
+												fill-rule="evenodd"
+												clip-rule="evenodd"
+												d="M64 5.103c-33.347 0-60.388 27.035-60.388 60.388 0 26.682 17.303 49.317 41.297 57.303 3.017.56 4.125-1.31 4.125-2.905 0-1.44-.056-6.197-.082-11.243-16.8 3.653-20.345-7.125-20.345-7.125-2.747-6.98-6.705-8.836-6.705-8.836-5.48-3.748.413-3.67.413-3.67 6.063.425 9.257 6.223 9.257 6.223 5.386 9.23 14.127 6.562 17.573 5.02.542-3.903 2.107-6.568 3.834-8.076-13.413-1.525-27.514-6.704-27.514-29.843 0-6.593 2.36-11.98 6.223-16.21-.628-1.52-2.695-7.662.584-15.98 0 0 5.07-1.623 16.61 6.19C53.7 35 58.867 34.327 64 34.304c5.13.023 10.3.694 15.127 2.033 11.526-7.813 16.59-6.19 16.59-6.19 3.287 8.317 1.22 14.46.593 15.98 3.872 4.23 6.215 9.617 6.215 16.21 0 23.194-14.127 28.3-27.574 29.796 2.167 1.874 4.097 5.55 4.097 11.183 0 8.08-.07 14.583-.07 16.572 0 1.607 1.088 3.49 4.148 2.897 23.98-7.994 41.263-30.622 41.263-57.294C124.388 32.14 97.35 5.104 64 5.104z"
+											/><path
+												d="M26.484 91.806c-.133.3-.605.39-1.035.185-.44-.196-.685-.605-.543-.906.13-.31.603-.395 1.04-.188.44.197.69.61.537.91zm2.446 2.729c-.287.267-.85.143-1.232-.28-.396-.42-.47-.983-.177-1.254.298-.266.844-.14 1.24.28.394.426.472.984.17 1.255zM31.312 98.012c-.37.258-.976.017-1.35-.52-.37-.538-.37-1.183.01-1.44.373-.258.97-.025 1.35.507.368.545.368 1.19-.01 1.452zm3.261 3.361c-.33.365-1.036.267-1.552-.23-.527-.487-.674-1.18-.343-1.544.336-.366 1.045-.264 1.564.23.527.486.686 1.18.333 1.543zm4.5 1.951c-.147.473-.825.688-1.51.486-.683-.207-1.13-.76-.99-1.238.14-.477.823-.7 1.512-.485.683.206 1.13.756.988 1.237zm4.943.361c.017.498-.563.91-1.28.92-.723.017-1.308-.387-1.315-.877 0-.503.568-.91 1.29-.924.717-.013 1.306.387 1.306.88zm4.598-.782c.086.485-.413.984-1.126 1.117-.7.13-1.35-.172-1.44-.653-.086-.498.422-.997 1.122-1.126.714-.123 1.354.17 1.444.663zm0 0"
+											/></g
+										>
+									</svg>
+								{/if}
+
+								{#if source.isSystemWide}
+									<svg
+										xmlns="http://www.w3.org/2000/svg"
+										class="h-10 w-10"
+										viewBox="0 0 24 24"
+										stroke-width="1.5"
+										stroke="currentColor"
+										fill="none"
+										stroke-linecap="round"
+										stroke-linejoin="round"
+									>
+										<path stroke="none" d="M0 0h24v24H0z" fill="none" />
+										<circle cx="12" cy="12" r="9" />
+										<line x1="3.6" y1="9" x2="20.4" y2="9" />
+										<line x1="3.6" y1="15" x2="20.4" y2="15" />
+										<path d="M11.5 3a17 17 0 0 0 0 18" />
+										<path d="M12.5 3a17 17 0 0 1 0 18" />
+									</svg>
+								{/if}
+							</div>
+							<div class="w-full flex flex-col">
+								<div class="h-10">
+									<h1 class="font-bold text-base truncate">{source.name}</h1>
+									{#if source.teams.length > 0 && source.teams[0]?.name}
+										<div class="truncate text-xs">{source.teams[0]?.name}</div>
+									{/if}
+								</div>
+								<div class="flex justify-end items-end space-x-2 h-10" />
+							</div>
+						</div>
+					</div>
+				</a>
+			{/each}
+		</div>
+	{/if}
